# plugins/api-platform-backend/docs/backend-openapi.yaml
openapi: 3.0.3
info:
  title: API Platform Backend
  description: Backend API for the API Platform plugin in Backstage
  version: 1.0.0
  contact:
    name: API Platform Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: /api/api-platform
    description: API Platform Backend

paths:
  /apis/definitions:
    get:
      summary: List API definitions
      description: Retrieve a paginated list of API definitions with optional search and ordering
      operationId: listApiDefinitions
      tags:
        - APIs
      parameters:
        - name: offset
          in: query
          description: Number of items to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Maximum number of items to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: orderBy
          in: query
          description: Field to order results by
          required: false
          schema:
            type: string
        - name: search
          in: query
          description: Search term to filter results
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of API definitions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiDefinition'
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /apis/count:
    get:
      summary: Get total count of APIs
      description: Retrieve the total number of API definitions
      operationId: getApisCount
      tags:
        - APIs
      responses:
        '200':
          description: Total count of APIs
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /apis/definitions/{system}/{apiName}:
    get:
      summary: Get API versions
      description: Retrieve all versions of a specific API
      operationId: getApiVersions
      tags:
        - APIs
      parameters:
        - name: system
          in: path
          required: true
          description: Code of the application
          schema:
            type: string
        - name: apiName
          in: path
          required: true
          description: Name of the API
          schema:
            type: string
      responses:
        '200':
          description: List of API versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiVersion'
        '404':
          description: API not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /apis/definitions/{system}/{apiName}/{apiVersion}:
    get:
      summary: Get specific API version
      description: Retrieve a specific version of an API
      operationId: getApiMatchingVersion
      tags:
        - APIs
      parameters:
        - name: system
          in: path
          required: true
          description: Code of the application
          schema:
            type: string
        - name: apiName
          in: path
          required: true
          description: Name of the API
          schema:
            type: string
        - name: apiVersion
          in: path
          required: true
          description: Version of the API
          schema:
            type: string
      responses:
        '200':
          description: API definition for the specified version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDefinition'
        '404':
          description: API or version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /catalog:
    post:
      summary: Register catalog entity
      description: Register a new entity in the catalog
      operationId: registerCatalogInfo
      tags:
        - Catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogEntity'
      responses:
        '201':
          description: Entity successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogEntity'
        '400':
          description: Invalid entity data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /catalog/{kind}/{name}:
    delete:
      summary: Unregister catalog entity
      description: Remove an entity from the catalog
      operationId: unregisterCatalogInfo
      tags:
        - Catalog
      parameters:
        - name: kind
          in: path
          required: true
          description: Kind of the entity (e.g., Component, API, System)
          schema:
            type: string
        - name: name
          in: path
          required: true
          description: Name of the entity
          schema:
            type: string
      responses:
        '204':
          description: Entity successfully unregistered
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /services:
    get:
      summary: List services
      description: Retrieve a list of all services
      operationId: listServices
      tags:
        - Services
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /services/{serviceName}:
    get:
      summary: Get service versions
      description: Retrieve all versions of a specific service
      operationId: getServiceVersions
      tags:
        - Services
      parameters:
        - name: serviceName
          in: path
          required: true
          description: Name of the service
          schema:
            type: string
      responses:
        '200':
          description: List of service versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceVersion'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /service-informations/{system}/{serviceName}/{serviceVersion}/{imageVersion}:
    get:
      summary: Get service information
      description: Retrieve detailed information about a specific service version
      operationId: getServiceInformation
      tags:
        - Service Information
      parameters:
        - name: system
          in: path
          required: true
          description: Application code
          schema:
            type: string
        - name: serviceName
          in: path
          required: true
          description: Name of the service
          schema:
            type: string
        - name: serviceVersion
          in: path
          required: true
          description: Version of the service
          schema:
            type: string
        - name: imageVersion
          in: path
          required: true
          description: Version of the image
          schema:
            type: string
      responses:
        '200':
          description: Service information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInformation'
        '404':
          description: Service information not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /service-informations:
    post:
      summary: Add service information
      description: Add new service information
      operationId: addServiceInformation
      tags:
        - Service Information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceInformation'
      responses:
        '201':
          description: Service information successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInformation'
        '400':
          description: Invalid service information data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /systems:
    get:
      summary: List systems
      description: Retrieve a list of all systems
      operationId: listSystems
      tags:
        - Systems
      responses:
        '200':
          description: List of systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/System'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /systems/{systemName}:
    get:
      summary: Get system
      description: Retrieve detailed information about a specific system
      operationId: getSystem
      tags:
        - Systems
      parameters:
        - name: systemName
          in: path
          required: true
          description: Name of the system
          schema:
            type: string
      responses:
        '200':
          description: System information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/System'
        '404':
          description: System not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    ApiDefinition:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the API
        name:
          type: string
          description: Name of the API
        version:
          type: string
          description: Version of the API
        description:
          type: string
          description: Description of the API
        spec:
          type: object
          description: OpenAPI specification
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - name
        - version

    ApiVersion:
      type: object
      properties:
        version:
          type: string
          description: Version identifier
        description:
          type: string
          description: Version description
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - version

    CatalogEntity:
      type: object
      properties:
        apiVersion:
          type: string
          description: API version of the entity
        kind:
          type: string
          description: Kind of entity (Component, API, System, etc.)
        metadata:
          type: object
          properties:
            name:
              type: string
            namespace:
              type: string
            title:
              type: string
            description:
              type: string
            annotations:
              type: object
              additionalProperties:
                type: string
            labels:
              type: object
              additionalProperties:
                type: string
          required:
            - name
        spec:
          type: object
          description: Entity specification
      required:
        - apiVersion
        - kind
        - metadata

    Service:
      type: object
      properties:
        name:
          type: string
          description: Name of the service
        description:
          type: string
          description: Description of the service
        versions:
          type: array
          items:
            type: string
          description: Available versions
        owner:
          type: string
          description: Owner of the service
      required:
        - name

    ServiceVersion:
      type: object
      properties:
        version:
          type: string
          description: Version identifier
        description:
          type: string
          description: Version description
        imageVersion:
          type: string
          description: Associated image version
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - version

    ServiceInformation:
      type: object
      properties:
        system:
          type: string
          description: Application code
        serviceName:
          type: string
          description: Name of the service
        serviceVersion:
          type: string
          description: Version of the service
        imageVersion:
          type: string
          description: Version of the image
        description:
          type: string
          description: Service description
        endpoints:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              method:
                type: string
              description:
                type: string
        dependencies:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
      required:
        - system
        - serviceName
        - serviceVersion
        - imageVersion

    System:
      type: object
      properties:
        name:
          type: string
          description: Name of the system
        description:
          type: string
          description: Description of the system
        owner:
          type: string
          description: Owner of the system
        components:
          type: array
          items:
            type: string
          description: Components belonging to the system
        domain:
          type: string
          description: Domain the system belongs to
      required:
        - name

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
      required:
        - error

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []

tags:
  - name: APIs
    description: API definition management
  - name: Catalog
    description: Catalog entity management
  - name: Services
    description: Service management
  - name: Service Information
    description: Detailed service information management
  - name: Systems
    description: System management