# plugins/api-platform-common/src/types.ts
import { Entity } from "@backstage/catalog-model";

export type ApiVersionDefinition = {
  entityRef: string;
  version: string;
  project: string;
};

export type ApiRelationServiceDefinition = {
  entityRef: string;
  version: string;
  lifecycle: string;
};

export type ApiRelationDefinition = {
  apiVersion: string;
  services: ApiRelationServiceDefinition[];
};

export type ServiceEnvironmentDefinition = {
  imageVersion: string;
  entityRef: string;
  platform: string;
};

export type ServiceVersionDefinition = {
  version: string;
  environments: {
    tst?: ServiceEnvironmentDefinition;
    gtu?: ServiceEnvironmentDefinition;
    uat?: ServiceEnvironmentDefinition;
    ptp?: ServiceEnvironmentDefinition;
    prd?: ServiceEnvironmentDefinition;
  };
};

export type ServiceDefinition = {
  name: string;
  serviceName: string;
  owner: string;
  system: string
  versions: ServiceVersionDefinition[];
};

export type ServiceApisDefinition = {
  consumedApis?: string[];
  providedApis?: string[];
};

export type SystemDefinition = {
  entity: Entity;
  apis: string[];
  services: string[];
};

export type ServiceInformation = {
  system: string;
  serviceName: string;
  serviceVersion: string;
  imageVersion: string;
  repository: string;
  sonarQubeProjectKey: string;
  apiDependencies: ServiceApisDefinition;
};


export type ApiDefinitionsListFields =
  | 'name'
  | 'description'
  | 'system';

export const APIDEFINITIONS_FIELDS = [
  'name',
  'description',
  'system'] as const;

export type ApiDefinitionListOptions = {
  offset?: number;
  limit?: number;
  orderBy?: {
    field: ApiDefinitionsListFields;
    direction: 'asc' | 'desc';
  };
  search?: string;
};

export type ApiDefinitionListResult = {
  items: Entity[];
  offset: number;
  limit: number;
  totalCount: number;
};